+++ Adj/Program/Primitive/Generation.hs
@@ -4,7 +4,7 @@
 module Adj.Program.Primitive.Generation where
 
 import Adj.Auxiliary (Casted, Casting ((=-), (-=)), type (=!?=), FG (FG), FFGH (FFGH), type (=!!??=), Structural (Structural))
-import Adj.Algebra.Category (Semigroupoid ((.)), Category ((.:), identity), Functor (map), Covariant, Bindable, Traversable, Identity (Identity), Day (Day), type (-->), type (-/->), Straight (Straight), Opposite, Kleisli (Kleisli), (->-), (->>-), (->>--), (->--), (-->--), (-/>/-), (-/>>/-), (=-=))
+import Adj.Algebra.Category (Semigroupoid ((.)), Category ((.:), identity), Functor (map), Covariant, Traversable, Identity (Identity), Day (Day), type (-->), type (-/->), Straight (Straight), Opposite, Kleisli (Kleisli), (->-), (->>-), (->>--), (->--), (-->--), (-/>/-), (-/>>/-), (=-=))
 import Adj.Algebra.Set (Setoid, (:*:) ((:*:)), (:+:) (This, That))
 
 newtype Generation f g o = Generation