+++ CHANGELOG.md
@@ -283,3 +283,4 @@
 # 0.1.4
 * Define `Generate` pattern
 * Remove `Component` typeclass
+* Change `Semimonoidal` type family definition